Rename AddToWishlistButton to WishlistButton
